Update patch set 2
I would prefer that you didn't submit this.

Sorry for the delay responding, some earlier review comments
inadvertently didn't get posted.   This issue is actually already
addressed as part of the larger OpenCORE 2.0 update that is about to be
merged (http://review.source.android.com/8604).  With the change, it
does not generate an error event when the timeout happens with data in
the jitterbuffer, and instead it attempts to playout what is left in
jitter buffer before sending EOS.

Patch-set: 2