Update patch set 1
Patch Set 1: No score

Regarding the EcnvBox, AVCSampleEntry, and AVCConfigurationBox: Even though EcnvBox parses the AVCConfigurationBox, is does not provide an AVCSampleEntry in the _pAVCSampleEntryVec array of the SampleDescriptionAtom object. As a result, the AVC related information does not get found for this track, when accessed through the SampleDescriptionAtom, e.g. SampleDescriptionAtom::getAVCNALLengthSize returns 0 as it does not find the _oAVC flag set and the corresponding entry in the _pAVCSampleEntryVec. 

With the change to SampleDescriptionAtom::SampleDescriptionAtom, parsing of EcnvBox results in adding an AVCSampleEntry object into the _pAVCSampleEntryVec and setting _oAVC=true just like it does after processing AVC_SAMPLE_ENTRY. This enables treating an encrypted AVC track the same way as a clear AVC track.

Perhaps there is a more elegant solution, but I don't see it.

Regarding iNumOutstandingInputBuffers initialization, I can add that.

I don't fully understand the workflow here, so please advise on what to do next.

Thanks.

Patch-set: 1