Update patch set 1
Patch Set 1: No score

Yes, but although it's a C++ wrapper, the parts of the AMR code (which is almost all files except the outermost packetvideo wrapper classes) that can be compiled as pure C still include this header (and probably need other parts of it).

I've made a standalone library which wraps the AMR codec from opencore (http://gitorious.org/opencore-amr/opencore-amr), and this is one of the few changes needed in order to compile those files as C, in order to minimize the dependencies of this library.

Patch-set: 1
Reviewer: Martin Storsjö <1002068@85c56323-6fa9-3386-8a01-6480fb634889>