Conditional defines OMX_CORE_LIBRARY.
OMX_CORE_LIBRARY is used to define the shared library that will be loaded
by omx core wrapper.  We don't want to copy the wrapper everywhere.
So, make it is defined in source code only when it is not defined by others.
Programmer can create a new wrapper with specified -DOMX_CORE_LIBRARY
options passed to compiler.