Update patch set 1
Patch Set 1: No score

Ah, yes, true, sorry for the confusion, NB_TRACK is 5.

However, is there any guarantee that each value of track between 0 and 4 is present twice? If a value is present only once, the corresponding element in the last half of indx isn't initialized.

I managed to set up a case where the value of the track variable was:
0, 0, 2, 3, 4, 0, 1, 2, 3, 4 (for each iteration through the loop)
In this case, indx[1+5] is never initialized properly, since track is 1 only once.

However; I'm not able to reproduce this situation later after fixing issue 10906 properly using negate, so I'm willing to drop this patch. But I don't understand this codebase well enough to be certain that there is no possibility for parts of indx to be left uninitialized.

Patch-set: 1
Reviewer: Martin Storsjö <1002068@85c56323-6fa9-3386-8a01-6480fb634889>