Update patch set 2
The array will always be initialized.

In set_sign.cpp, vector ipos[] is initialized with values between 0 and 4, in an order that depends on the analyzed data.

In s10_8pf.cpp, the values of ipos[] are used as initial values to fill vector index[], then vector codvec[]. In the for-loops, the values of ipos[] are increased by step ( step==STEP defined as 5).

As the values of variable 'track', on function build_code, on file c1035pf.cpp, are equal to the module of codvec by 5, no matter which values they are, 'track' will have the same  values as ipos[] had on set_sign.cpp. 

track = ipos[] + 5*k, where k depends on peak location

track = module((ipos[] + 5*k), 5) = ipos[]

Therefore, the range 0 to 4 is always guaranteed.

As 'track' is used to initialized indx[], and 'track' goes between 0 and 4, initialization for the whole array will be completed, as explain earlier.

No need to over-extend initialization

Patch-set: 2