Fix issue 3101 - No interleaving to mp4 files for just one track
The patch efficiently reduces latency when playing pure video files
while they are recorded. Tests showed that pure video contents are
ok, files with audio and video are unchanged.

This patch has been submitted and got merged ealier, but it was
lost probably with the drop from Opencore2. It is suggested to
backport it to the Opencore sources to avoid loosing it again.