Update patch set 2
I found your modifications now.

The reason why I was looking for the patch is that Opencore as part of ADP1.6 is still writing video files in chunks of 1s intervals. With my original patch this was fixed. Do you have an explanation why this happens? Did you actually test your new code for replacing the original patch?

Patch-set: 2