Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

ISO/IEC 144496-12_2005:AMD1_2007
The pixel aspect ratio and clean aperture of the video may be specified using the ‘pasp’ and ‘clap’ sample
entry boxes, respectively. These are both optional; if present, they over-ride the declarations (if any) in
structures specific to the video codec, which structures should be examined if these boxes are absent.


The "transformation matrix" is an additional way to manipulate the video stream. A correct player behavior is to first look for PAR or SAR, and scale the video accordingly. After this is done, the changes by the "transformation matrix" have to be applied.

The "transformation matrix" isn't used in all ISO base media file format (ISO/IEC 14496-12) based formats. It is used in MOV(ftyp "QT ") and M4V(ftyp "M4V "), as far as I know.

ISO/IEC 14496-12_2005
Matrix values which occur in the headers specify a transformation of video images for presentation. Not all
derived specifications use matrices; if they are not used, they shall be set to the identity matrix, If a matrix is
used, the point (p,q) is transformed into (p', q') using the matrix as follows:
In the MP4 file format(ISO/IEC 14496-14) the "transformation matrix" isn't used. In the MP4 file format the same could be done via BIFS.

ISO/IEC 14496-14_2003
Composition
In MPEG-4 both visual and aural composition are done using the BIFS system. Therefore structures marked as “template” in the ISO Base Media Format which pertain to composition, including fields such as matrices, layers, graphics modes (and their opcolors), volumes, and balance values, from the MovieHeaderBox and TrackHeaderBox, are all set to their default values in the file format. These fields do not define visual or audio composition in MPEG-4; in MPEG-4, the BIFS system defines the composition.

Patch-set: 1
Reviewer: Chen Guoyin <1002702@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1