Update patch set 1
Patch Set 1: No score

This is just a enhancement for mp4 parser to support some clips which may not strictly fellow mp4 spec. For esd information, it can be constructed from other part of mp4 box.It still can be decoded by video decoder if we construct this. But it is ok if we dont have this enhancement.

Patch-set: 1
Reviewer: Chen Guoyin <1002702@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0