Update patch set 1
Most of our commit messages are prefixed by RIO-# referencing PV's internal source control tree.  RIO-7937 was integrated and tested last week and should be pushed as part of our 2.06 release within a few days.

I'll add an update here with the URL to the commit in gitweb when this happens.

Patch-set: 1