Update patch set 1
Patch Set 1:

I'm not sure what to expect here.

Process-wise, uploading this change against froyo was the right thing to do as froyo is closer to Google's internal master than the open-source master - thanks for noticing and doing the right thing here.

However, my understanding is that we're not using this specific code internally (and in fact it's very possible that this is even dead code in froyo as far as we're concerned, though things might be different on other builds). It's unclear that we'll ever want to cherry-pick this change into our internal master.

Patch-set: 1